Education, Fonts, Images, Videos
Added the Education section included activities. Fixed the head/font imports. De-hard coded a part of the header that was making education and archive/resources too close together. Added images for the cards and activities, as well as video demos of projects.  Added contact info with tags to the name container.